requirements.txt

folium
fpdf
geocoder
geopy
gitpython
googlemaps
gunicorn
h5py
keras
markdown
matplotlib
ml-dtypes
numpy
opencv-python
pandas
pillow
protobuf
pyarrow
pydeck
pyjwt
python-dotenv
requests
scikit-learn
seaborn
streamlit
tensorflow
uvicorn
streamlit-option-menu==0.3.13
streamlit-lottie==0.0.5